+++ src/net/net-tcp-rpc-ext-uniq-bloom.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void uniq_bloom_clear_slot (int sid);
+int uniq_bloom_test_and_set (int sid, unsigned ip, const unsigned char *ipv6);